backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy
databases[aiosqlite]
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0.0
bcrypt==3.2.0
passlib==1.7.4
email-validator==2.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
alembic==1.11.1
aiosqlite==0.19.0
pytest-cov==4.1.0
coverage==7.3.2
pytest-mock==3.10.0
Faker==18.9.0
python-dotenv==1.0.0
pytest-xdist==3.4.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
twilio==8.13.0
vonage==3.0.0
python-dateutil==2.8.2
aiofiles==23.2.1
jinja2==3.1.2
werkzeug


